# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Management API
  description: API REST pour la gestion de plantes domestiques
  version: 1.0.0
  contact:
    name: Plant Team
servers:
  - url: http://localhost:8080/api/v1
    description: Development (Quarkus local)
  # a remplacer par l'URL du cluster Kubernetes université ?
  - url: http://plant-backend.etu-k8s.lan/api/v1
    description: Production (Kubernetes cluster)
...
security:
  - bearerAuth: []
paths:
...
  /houses:
    post:
      tags: [House]
      summary: Créer une maison
      description: L'utilisateur devient OWNER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseRequest'
      responses:
        '201':
          description: Maison créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseResponse'
        '400':
          description: Utilisateur a déjà une maison
  /houses/join:
    post:
      tags: [House]
      summary: Rejoindre une maison
      description: L'utilisateur devient MEMBER
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode]
              properties:
                inviteCode:
                  type: string
                  example: "ABC12345"
      responses:
        '200':
          description: Maison rejointe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseResponse'
        '400':
          description: Utilisateur a déjà une maison
        '404':
          description: Code invalide
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mcpApiKey:
      type: apiKey
      in: header
      name: X-MCP-API-Key
      description: API Key pour l'accès MCP (Goose/Mistral)
...
  schemas:
...
    CreateHouseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
    HouseResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
          format: date-time
...